[Cython] Fix a few compiler warnings from Eigen

+++ include/cantera/cython/kinetics_utils.h
@@ -42,13 +42,13 @@ inline void sparseCscData(const Eigen::SparseMatrix<double>& mat,
 
     const double* valuePtr = mat.valuePtr();
     const int* innerPtr = mat.innerIndexPtr();
-    for (size_t i = 0; i < mat.nonZeros(); ++i) {
+    for (int i = 0; i < mat.nonZeros(); ++i) {
         value[i] = valuePtr[i];
         inner[i] = innerPtr[i];
     }
 
     const int* outerPtr = mat.outerIndexPtr();
-    for (size_t i = 0; i < mat.outerSize() + 1; ++i) {
+    for (int i = 0; i < mat.outerSize() + 1; ++i) {
         outer[i] = outerPtr[i];
     }
 }
